requirements-socketio.txt
python-socketio[asyncio]==5.10.0
aiohttp==3.9.1
aioredis==2.0.1
gunicorn==21.2.0
